Code Refactor
- Refactor front category state to include subviews when loading the
category's information
