+++ public/js/categories/CategoriesHeaderCtrl.js
