+++ public/js/categories/CategoriesModule.js
@@ -2,9 +2,9 @@
 
 define( function ( require ) {
     var CategoriesBaseCtrl      = require( 'categories/CategoriesBaseCtrl' );
-    var CategoriesContentCtrl   = require( 'categories/CategoriesContentCtrl' );
     var CategoriesCreateCtrl    = require( 'categories/CategoriesCreateCtrl' );
     var CategoriesEditCtrl      = require( 'categories/CategoriesEditCtrl' );
+    var CategoriesHeaderCtrl    = require( 'categories/CategoriesHeaderCtrl' );
     var CategoriesListCtrl      = require( 'categories/CategoriesListCtrl' );
     var CategoriesRouter        = require( 'categories/CategoriesRouter' );
     var CategoriesService       = require( 'categories/CategoriesService' );
@@ -15,11 +15,11 @@ define( function ( require ) {
 
     CategoriesModule.controller( 'CategoriesBaseCtrl', [ '$scope', CategoriesBaseCtrl ]);
 
-    CategoriesModule.controller( 'CategoriesContentCtrl', [ '$scope', '$stateParams', 'CategoriesService', CategoriesContentCtrl ]);
-
     CategoriesModule.controller( 'CategoriesCreateCtrl', [ '$scope', 'events', 'CategoriesService', CategoriesCreateCtrl ]);
 
     CategoriesModule.controller( 'CategoriesEditCtrl', [ '$scope', '$stateParams', 'events', 'CategoriesService', CategoriesEditCtrl ]);
+
+    CategoriesModule.controller( 'CategoriesHeaderCtrl', [ '$scope', '$stateParams', 'CategoriesService', CategoriesHeaderCtrl ]);
 
     CategoriesModule.controller( 'CategoriesListCtrl', [ '$scope', 'CategoriesService', CategoriesListCtrl ]);
