+++ views/partials/categories/front.jade
@@ -1,0 +1,1 @@
+section#category-header( ui-view="category-header" )
